fix: divide data category into a seperate front end table section
